fix: removing duplicated code
