+++ apps/meteor/app/api/server/api.ts
@@ -3,9 +3,7 @@ import { Logger } from '@rocket.chat/logger';
 import { Users } from '@rocket.chat/models';
 import { Random } from '@rocket.chat/random';
 import type { JoinPathPattern, Method } from '@rocket.chat/rest-typings';
-import { ajv } from '@rocket.chat/rest-typings/src/v1/Ajv';
 import { wrapExceptions } from '@rocket.chat/tools';
-import type { ValidateFunction } from 'ajv';
 import express from 'express';
 import type { Request, Response } from 'express';
 import { Accounts } from 'meteor/accounts-base';
@@ -31,8 +29,6 @@ import type {
 	PartialThis,
 	SuccessResult,
 	TypedThis,
-	TypedAction,
-	TypedOptions,
 	UnauthorizedResult,
 } from './definition';
 import { getUserInfo } from './helpers/getUserInfo';
@@ -145,14 +141,7 @@ const generateConnection = (
 	clientAddress: ipAddress,
 });
 
-export class APIClass<
-	TBasePath extends string = '',
-	TOperations extends {
-		[x: string]: unknown;
-	} = {},
-> {
-	public typedRoutes: Record<string, Record<string, unknown>> = {};
-
+export class APIClass<TBasePath extends string = ''> {
 	protected apiPath?: string;
 
 	readonly version?: string;
@@ -507,182 +496,6 @@ export class APIClass<
 		return routeActions.map((action) => this.getFullRouteName(route, action));
 	}
 
-	private registerTypedRoutesLegacy<TSubPathPattern extends string, TOptions extends Options>(
-		method: Method,
-		subpath: TSubPathPattern,
-		options: TOptions,
-	): void {
-		const { authRequired, validateParams } = options;
-
-		const opt = {
-			authRequired,
-			...(validateParams &&
-				method.toLowerCase() === 'get' &&
-				('GET' in validateParams
-					? { query: validateParams.GET }
-					: {
-							query: validateParams as ValidateFunction<any>,
-						})),
-
-			...(validateParams &&
-				method.toLowerCase() === 'post' &&
-				('POST' in validateParams ? { query: validateParams.POST } : { body: validateParams as ValidateFunction<any> })),
-
-			...(validateParams &&
-				method.toLowerCase() === 'put' &&
-				('PUT' in validateParams ? { query: validateParams.PUT } : { body: validateParams as ValidateFunction<any> })),
-			...(validateParams &&
-				method.toLowerCase() === 'delete' &&
-				('DELETE' in validateParams ? { query: validateParams.DELETE } : { body: validateParams as ValidateFunction<any> })),
-
-			tags: ['Missing Documentation'],
-			response: {
-				200: ajv.compile({
-					type: 'object',
-					properties: {
-						success: { type: 'boolean' },
-						error: { type: 'string' },
-					},
-					required: ['success'],
-				}),
-			},
-		};
-
-		this.registerTypedRoutes(method, subpath, opt);
-	}
-
-	private registerTypedRoutes<
-		TSubPathPattern extends string,
-		TOptions extends TypedOptions,
-		TPathPattern extends `${TBasePath}/${TSubPathPattern}`,
-	>(method: Method, subpath: TSubPathPattern, options: TOptions): void {
-		const path = `/${this.apiPath}/${subpath}`.replaceAll('//', '/') as TPathPattern;
-		this.typedRoutes = this.typedRoutes || {};
-		this.typedRoutes[path] = this.typedRoutes[subpath] || {};
-		const { query, authRequired, response, body, tags, ...rest } = options;
-		this.typedRoutes[path][method.toLowerCase()] = {
-			...(response && {
-				responses: Object.fromEntries(
-					Object.entries(response).map(([status, schema]) => [
-						status,
-						{
-							description: '',
-							content: {
-								'application/json': 'schema' in schema ? { schema: schema.schema } : schema,
-							},
-						},
-					]),
-				),
-			}),
-			...(query && {
-				parameters: [
-					{
-						schema: query.schema,
-						in: 'query',
-						name: 'query',
-						required: true,
-					},
-				],
-			}),
-			...(body && {
-				requestBody: {
-					required: true,
-					content: {
-						'application/json': { schema: body.schema },
-					},
-				},
-			}),
-			...(authRequired && {
-				...rest,
-				security: [
-					{
-						userId: [],
-						authToken: [],
-					},
-				],
-			}),
-			tags,
-		};
-	}
-
-	private method<TSubPathPattern extends string, TOptions extends TypedOptions, TPathPattern extends `${TBasePath}/${TSubPathPattern}`>(
-		method: Method,
-		subpath: TSubPathPattern,
-		options: TOptions,
-		action: TypedAction<TOptions>,
-	): APIClass<
-		TBasePath,
-		| TOperations
-		| ({
-				method: Method;
-				path: TPathPattern;
-		  } & Omit<TOptions, 'response'>)
-	> {
-		this.addRoute([subpath], { ...options }, { [method.toLowerCase()]: { action } } as any);
-		this.registerTypedRoutes(method, subpath, options);
-		return this;
-	}
-
-	get<TSubPathPattern extends string, TOptions extends TypedOptions, TPathPattern extends `${TBasePath}/${TSubPathPattern}`>(
-		subpath: TSubPathPattern,
-		options: TOptions,
-		action: TypedAction<TOptions>,
-	): APIClass<
-		TBasePath,
-		| TOperations
-		| ({
-				method: 'GET';
-				path: TPathPattern;
-		  } & Omit<TOptions, 'response'>)
-	> {
-		return this.method('GET', subpath, options, action);
-	}
-
-	post<TSubPathPattern extends string, TOptions extends TypedOptions, TPathPattern extends `${TBasePath}/${TSubPathPattern}`>(
-		subpath: TSubPathPattern,
-		options: TOptions,
-		action: TypedAction<TOptions>,
-	): APIClass<
-		TBasePath,
-		| TOperations
-		| ({
-				method: 'POST';
-				path: TPathPattern;
-		  } & Omit<TOptions, 'response'>)
-	> {
-		return this.method('POST', subpath, options, action);
-	}
-
-	put<TSubPathPattern extends string, TOptions extends TypedOptions, TPathPattern extends `${TBasePath}/${TSubPathPattern}`>(
-		subpath: TSubPathPattern,
-		options: TOptions,
-		action: TypedAction<TOptions>,
-	): APIClass<
-		TBasePath,
-		| TOperations
-		| ({
-				method: 'PUT';
-				path: TPathPattern;
-		  } & Omit<TOptions, 'response'>)
-	> {
-		return this.method('PUT', subpath, options, action);
-	}
-
-	delete<TSubPathPattern extends string, TOptions extends TypedOptions, TPathPattern extends `${TBasePath}/${TSubPathPattern}`>(
-		subpath: TSubPathPattern,
-		options: TOptions,
-		action: TypedAction<TOptions>,
-	): APIClass<
-		TBasePath,
-		| TOperations
-		| ({
-				method: 'DELETE';
-				path: TPathPattern;
-		  } & Omit<TOptions, 'response'>)
-	> {
-		return this.method('DELETE', subpath, options, action);
-	}
-
 	addRoute<TSubPathPattern extends string>(
 		subpath: TSubPathPattern,
 		operations: Operations<JoinPathPattern<TBasePath, TSubPathPattern>>,
@@ -895,7 +708,7 @@ export class APIClass<
 					endpoints: operations[method as keyof Operations<TPathPattern, TOptions>] as unknown as Record<string, string>,
 				});
 
-				this.registerTypedRoutesLegacy(method as Method, route, {
+				this.router.registerTypedRoutesLegacy(method as Method, route, {
 					...options,
 					...operations[method as keyof Operations<TPathPattern, TOptions>],
 				});
