+++ apps/meteor/app/api/server/v1/custom-sounds.ts
@@ -6,7 +6,7 @@ import { escapeRegExp } from '@rocket.chat/string-helpers';
 import { API } from '../api';
 import { getPaginationItems } from '../helpers/getPaginationItems';
 
-API.v1.get(
+API.v1.router.get(
 	'custom-sounds.list',
 	{
 		response: {
