+++ apps/meteor/app/api/server/v1/invites.ts
@@ -15,7 +15,7 @@ import { useInviteToken } from '../../../invites/server/functions/useInviteToken
 import { validateInviteToken } from '../../../invites/server/functions/validateInviteToken';
 import { API } from '../api';
 
-API.v1
+API.v1.router
 	.get(
 		'listInvites',
 		{
